+++ src/main/java/shadows/apotheosis/mixin/ItemMixin.java
@@ -4,6 +4,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 
 import net.minecraft.world.item.Item;
+import shadows.apotheosis.Apotheosis;
 import shadows.apotheosis.ench.table.IEnchantableItem;
 
 @Mixin(Item.class)
@@ -16,7 +17,7 @@ public class ItemMixin implements IEnchantableItem {
 	 */
 	@Overwrite
 	public int getEnchantmentValue() {
-		return 1;
+		return Apotheosis.enableEnch ? 1 : 0;
 	}
 
 }